requirements.txt
flet
pyserial
evdev
